+++ CHANGELOG.md
@@ -28,10 +28,12 @@
 - [The user may set description and labels of an Enso Cloud asset
   programmatically.][11255]
 - [DB_Table may be saved as a Data Link.][11371]
+- [Support for dates before 1900 in Excel and signed AWS requests.][11373]
 
 [11235]: https://github.com/enso-org/enso/pull/11235
 [11255]: https://github.com/enso-org/enso/pull/11255
 [11371]: https://github.com/enso-org/enso/pull/11371
+[11373]: https://github.com/enso-org/enso/pull/11373
 
 #### Enso Language & Runtime